fix: address PR review feedback
- Split configureProcessGroup into OS-specific files with build tags
  (processgroup_unix.go, processgroup_windows.go) for cross-platform builds
- Add nil-safety comment on cmd.Cancel documenting the Go exec invariant
- Bump commandTimeout from 10s to 30s for slow CI runners
- Filed #277 for io.Reader refactor of wizard functions
